refactor: 검색 시 타이틀 및 에러 메세지를 컨트롤하는 메서드로 분리 후 onSearch에 전달
